Refactor each resource factory to separate file
