+++ tests/factory_fixtures/__init__.py
@@ -1,1 +1,4 @@
 from .lease import *
+from .tenant import *
+from .user import *
+from .property import *
